Several compiler warnings resolved

+++ IsometricEngine/src/PathfinderAgent.h
@@ -5,10 +5,17 @@
 
 class PathfinderAgent : public Agent
 {
-public:
-	PathfinderAgent(int initialPosX, int initialPosY, sf::String unit) :Agent(initialPosX, initialPosY, -1, -1, -1, -1, true, unit) { isPathfinderAgent = true; };
+   public:
+	PathfinderAgent(float initialPosX, float initialPosY, sf::String unit)
+		: Agent(initialPosX, initialPosY, -1, -1, -1, -1, true, unit)
+	{
+		isPathfinderAgent = true;
+	};
 
-	void setStartingCell(BattlefieldCell* startingCell) { current = startingCell; }
+	void setStartingCell(BattlefieldCell* startingCell)
+	{
+		current = startingCell;
+	}
 	void setPath(std::vector<BattlefieldCell*>* TargetPath);
 	void update();
